Add compute_order function
Compute the order of convergence

The sequence {x_i} is generated by fixed-point iteration. If sequence
converege to L and error is denoted by e_i = |x_i-L|/|L|.

As i tend to infinty, the order of converge alpha could be approximated by

	alpha = log(|e_i/e_{i-1}|)/log(|e_{i-1}/e_{i-2}|)

And the fixed-point method is linearly convergent.